OpenVINO: updated grasp detector to use base Classifier
This patch updated the "GraspDetector" to use the base Classifier.
The base Classifier was created as a shared pointer, which will be
released by system. Thus no need to explitly release "classifier_"
in the destructor "~GraspDetector()".

The base Classifer will create the right classifier (Caffe or OpenVINO)
based on build option.

The Caffe Classifer is modified to be derived from the base Classifer.

Signed-off-by: Sharron LIU <sharron.liu@intel.com>