Adding support for python bindings
Can build shared library for python bindings by specifying
-DBUILD_PYTHON=ON when calling cmake. This requires pybind11
to be cloned into build's parent directory. The python bindings
expose a very minimal interface: it allows construction of a
`GraspDetector` and calling `detectGrasps` which accepts a
previously instantiated `GraspDetector` and a point cloud as a
numpy array (of shape (n, 3)). This function returns a numpy
array of shape (number of grasps, 14) where the columns specify
the position, orientation (rotation matrix in row-major order), the width
of the gripper, and the score of the grasp.